Split spatial example
